+++ spec/ajax-datatables-rails/datatable/column_spec.rb
@@ -128,6 +128,14 @@ RSpec.describe AjaxDatatablesRails::Datatable::Column do
     end
   end
 
+  describe 'unsearchable column' do
+    let(:column) { datatable.datatable.columns.find{ |c| c.data == 'email_hash' } }
+
+    it 'is not searchable' do
+      expect(column.searchable?).to eql(false)
+    end
+  end
+
   describe '#formatter' do
     let(:datatable) { DatatableWithFormater.new(sample_params) }
     let(:column) { datatable.datatable.columns.find { |c| c.data == 'last_name' } }