+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* Implementing `searchable: false` tests
+
 ## 1.4.0 (2022-12-18)
 
 * Improve tests